change(Mesa): Cambio de nombre de un atirbuto de la entidad Mesa y agregando uno nuevo - MesaService: Cambio del crud para la utilizacion de dtos - MesaRequestDTO y MesaResponseDTO: Nuevos DTOs para la mesa. - MesaController: Cambiando los metodos con buenas practicar y agregando seguridad (permisos) a los enpoints. - SecurityConfi: Agregamos los enpoints de la mesa para definir que tipo de Rol esta permito para utilizar el enpoint.

+++ src/main/java/com/backend/sistemarestaurante/modules/mesas/MesaService.java
@@ -1,10 +1,15 @@
 package com.backend.sistemarestaurante.modules.mesas;
 
+import com.backend.sistemarestaurante.modules.mesas.dto.MesaRequestDTO;
+import com.backend.sistemarestaurante.modules.mesas.dto.MesaResponseDTO;
+import com.backend.sistemarestaurante.shared.exceptions.ResourceNotFoundException;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class MesaService {
@@ -13,29 +18,68 @@ public class MesaService {
     @Autowired
     private MesaRepository mesaRepository;
 
+    @Autowired
+    private ModelMapper modelMapper;
+
     //metodo listar mesas
-    public List<Mesa> getAll() {
-        return (List<Mesa>) mesaRepository.findAll();
+    public List<MesaResponseDTO> getAll() {
+        // Obtener todos las mesas registradas en db
+        List<Mesa> mesas = (List<Mesa>) mesaRepository.findAll();
+
+        // Convetir la lista de entidades a lista de DTOs
+        return mesas.stream()
+                .map(mesa -> modelMapper.map(mesa, MesaResponseDTO.class))
+                .collect(Collectors.toList());
     }
 
     // Metodo buscar mesa por id
-    public Optional<Mesa> getById(Long id) {
-        return mesaRepository.findById(id);
+    public MesaResponseDTO getById(Long id) {
+        Mesa mesa = mesaRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Mesa no encontrada con el id: " + id));
+
+        // Convertir a DTO
+        return modelMapper.map(mesa, MesaResponseDTO.class);
+
     }
 
     // Metodo crear Mesa
-    public Mesa create(Mesa mesa) {
-        return mesaRepository.save(mesa);
+    public MesaResponseDTO create(MesaRequestDTO mesaRequestDTO) {
+        // Convertir el dto a entidad
+        Mesa nuevaMesa = modelMapper.map(mesaRequestDTO, Mesa.class);
+
+        // Guardar el objecto convertido a la base de datos
+        nuevaMesa = mesaRepository.save(nuevaMesa);
+
+        // Devolver el DTO de respuesta
+        return modelMapper.map(nuevaMesa, MesaResponseDTO.class);
     }
 
     // Metodo actualizar Mesa
-    public Mesa update(Mesa mesa){
-        //Mesa mesaFormOb = getById(mesa.getId()).orElseThrow(ResourceNotFoundException::new);
-        return null;
+    public MesaResponseDTO update(Long id, MesaRequestDTO mesaRequestDTO){
+        
+        // Verificar si la mesa exista
+        Mesa mesaExistente = mesaRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Mesa no encontrada con el id: " + id));
+        
+        // Copiar propiedades del DTO a la entidad (ignorando el ID)
+        modelMapper.map(mesaRequestDTO, mesaExistente);
+
+        // Guardar (usando la variable correcta: categoriaExistente)
+        Mesa mesaActualizada = mesaRepository.save(mesaExistente);
+
+        // Convertir a DTO de respuesta y retornar
+        return modelMapper.map(mesaActualizada, MesaResponseDTO.class);
     }
 
     // Metodo eliminar Mesa
-    public void delete(Long id) {
+    public void eliminarMesa(Long id) {
+        
+        // Verifiacr si existe la mesa que desea eliminar
+        if (!mesaRepository.existsById(id)){
+            throw new ResourceNotFoundException("Mesa no encontrada con id: " + id);
+        }
+        
+        // Eliminacion
         mesaRepository.deleteById(id);
     }
